Initialize npm-yarn-scanner project
Set up the project with an initial index.js file. This lays the foundation
for future development of the npm-yarn-scanner tool, which will analyze
and compare npm and yarn dependencies in JavaScript projects. The initial
commit establishes the basic structure and prepares the repository for
subsequent feature additions and improvements.

It should be more customisable to add packages but getting this out for now